chore(example): make merkle and bin and mod
