+++ examples/cli.rs
@@ -6,13 +6,9 @@ use std::num::NonZeroUsize;
 
 use clap::{Parser, ValueEnum};
 use halo2_proofs::halo2curves;
-
-use merkle::MerkleTreeUpdateCircuit;
-
 use poseidon::poseidon_step_circuit::TestPoseidonCircuit;
 use sirius::{
     ff::{FromUniformBytes, PrimeField, PrimeFieldBits},
-    gadgets::merkle_tree::off_circuit::Tree,
     ivc::{step_circuit::trivial, CircuitPublicParamsInput, PublicParams, StepCircuit, IVC},
     poseidon::ROPair,
 };
@@ -24,6 +20,8 @@ mod poseidon;
 
 #[allow(dead_code)]
 mod merkle;
+
+use merkle::{MerkleTreeUpdateCircuit, Tree};
 
 #[derive(Parser, Debug)]
 #[command(version, about, long_about = None)]
