+++ examples/merkle/mod.rs
@@ -1,0 +1,5 @@
+pub mod circuit;
+pub mod merkle_tree_gadget;
+
+pub use circuit::MerkleTreeUpdateCircuit;
+pub use merkle_tree_gadget::off_circuit::Tree;
